refactored header to have abstracted, data driven components. And restructured atomic design folders
