+++ package.json
@@ -52,6 +52,7 @@
     "firebase": "^7.14.2",
     "firebaseui": "^4.5.0",
     "node-sass": "^4.14.0",
+    "prop-types": "^15.7.2",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-redux": "^7.2.0",
