+++ src/components/organisms/Header/nav-links.js
@@ -1,0 +1,12 @@
+const navLinks = [
+  {
+    name: 'Home',
+    linkto: '/',
+  },
+  {
+    name: 'Prompts',
+    linkto: '/prompts',
+  },
+];
+
+export default navLinks;
