+++ src/components/pages/PageHome.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Header from '../organisms/Header';
+import Header from '../organisms/Header/Header';
 
 function Home() {
   return (
